Storybook: ReferenceArea - simplify and follow pattern
