+++ storybook/stories/API/cartesian/ReferenceArea.mdx
@@ -1,7 +1,13 @@
-import { ArgTypes, Meta, Stories } from '@storybook/blocks';
+import { Meta, Story, Canvas, Controls } from '@storybook/blocks';
 import * as ComponentStories from './ReferenceArea.stories';
 
 # ReferenceArea
+
+<Meta of={ComponentStories} />
+
+<Canvas>
+  <Story name="API" of={ComponentStories} />
+</Canvas>
 
 ## Parent Component
 The ReferenceArea can be used within the following parent components:
@@ -15,7 +21,4 @@ The ReferenceArea can be used with the following child components: `<Label/>`
 
 Properties in the groups Other and Internal are not recommended to be used.
 
-<Meta of={ComponentStories} />;
-<ArgTypes of={ComponentStories} sort={'requiredFirst'} />;
-<Stories />;
-
+<Controls of={ComponentStories.API} />
